+++ src/reactionsystem.jl
@@ -142,6 +142,7 @@ function get_netstoich(subs, prods, sstoich, pstoich)
     ns
 end
 
+
 """
 $(TYPEDEF)
 
@@ -187,11 +188,11 @@ struct ReactionSystem{U <: Union{Nothing,MT.AbstractSystem}} <: MT.AbstractTimeD
     """Non-`Reaction` equations that further constrain the system"""
     constraints::U
 
-    function ReactionSystem(eqs, iv, states, ps, observed, name, systems, defaults, connection_type, csys; checks::Bool = true)
-
-        iv′ = value(iv)
-        states′ = value.(states)
-        ps′ = value.(ps)
+    function ReactionSystem(eqs, iv, states, ps, observed, name, systems, defaults, connection_type, csys; 
+                            checks::Bool=true, skipvalue=false)
+        iv′     = value(iv)        
+        states′ = skipvalue ? states : value.(MT.scalarize(states))
+        ps′     = skipvalue ? ps : value.(MT.scalarize(ps))
 
         if checks
             check_variables(states′, iv′)
@@ -200,7 +201,6 @@ struct ReactionSystem{U <: Union{Nothing,MT.AbstractSystem}} <: MT.AbstractTimeD
         end
         rs = new{typeof(csys)}(collect(eqs), iv′, states′, ps′, observed, name, systems, defaults, connection_type, csys)
         checks && validate(rs)
-
         rs
     end
 end
@@ -214,12 +214,37 @@ function ReactionSystem(eqs, iv, species, ps;
                         defaults=_merge(Dict(default_u0), Dict(default_p)),
                         connection_type=nothing,
                         checks = true, 
-                        constraints = nothing)
+                        constraints = nothing,
+                        skipvalue = false)
     name === nothing && throw(ArgumentError("The `name` keyword must be provided. Please consider using the `@named` macro"))
 
-    ReactionSystem(eqs, iv, species, ps, observed, name, systems, 
-                   defaults, connection_type, constraints; checks = checks)
+    ReactionSystem(eqs, iv, species, ps, observed, name, systems, defaults, connection_type, 
+                   constraints; checks = checks, skipvalue = skipvalue)
 end
+
+function ReactionSystem(rxs::Vector{<:Reaction}, iv; kwargs...)  
+    t   = value(iv)   
+    sts = Set(spec for rx in rxs for spec in rx.substrates)
+    foreach(v -> push!(sts,v), (prod for rx in rxs for prod in rx.products))
+
+    ps   = Set()
+    vars = Set()
+    for rx in rxs
+        MT.get_variables!(vars, rx.rate)
+        for var in vars
+            isequal(t,var) && continue
+            if MT.isparameter(var) 
+                push!(ps, var)
+            else
+                push!(sts, var)
+            end
+        end
+        empty!(vars)
+    end
+
+    ReactionSystem(rxs, t, collect(sts), collect(ps); skipvalue=true, kwargs...)
+end
+
 
 function ReactionSystem(iv; kwargs...)
     ReactionSystem(Reaction[], iv, [], []; kwargs...)